Derive examples path from current file
This changes the behaviour of TestSolidPython.expand_scad_path so
that it finds the scad directory by going up one directory from
solid/test.

This allows the tests to be run with run_all_tests.sh or tox. Previously
running the tests with tox was failing due to not finding the example
files.