Write docstrings derived from OpenSCAD manual
This change adds docstrings which are suitable for sphinx autodoc
generation.

The OpenSCAD User Manual is licensed under the
Creative Commons Attribution-ShareAlike License, and my research
suggests that it is safe to mix this with LGPL code.

These docs can be generated with

  tox -edocs

It would be very useful now for these docs to be hosted on
https://readthedocs.org/