Stopgap change so we depend on the euclid3 module (https://pypi.python.org/pypi/euclid3/0.01), not the original PyEuclid.  Works for Py2 & Py3
This is unfortunate since the code in the current master repo of PyEuclid at https://github.com/ezag/pyeuclid has been Python3-compatible for at least a year.  However, the original PyPI maintainer of PyEuclid is unresponsive, so ezag can't update the PyPI dependency as of December 2016.

Optimally, he'll be able to take over the original package name at some point in the future, release a new version on PyPI and we can go back to the original dependency.  Until then, this should serve